[tests] moved AbstractProjectTestCase to testsupport bundle
