+++ plugins/org.locationtech.udig.catalog.tests.geotiff/META-INF/MANIFEST.MF
@@ -3,7 +3,6 @@ Bundle-ManifestVersion: 2
 Bundle-Name: Geotiff Test Plug-in
 Bundle-SymbolicName: org.locationtech.udig.catalog.test.geotiff;singleton:=true
 Bundle-Version: 2.0.0.qualifier
-Bundle-Activator: org.locationtech.udig.catalog.tests.internal.geotiff.GeoTiffPlugin
 Bundle-Vendor: udig.refractions.net
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
