--- plugins/org.locationtech.udig.catalog.tests.geotiff/src/org/locationtech/udig/catalog/tests/internal/geotiff/GeoTiffPlugin.java
@@ -1,64 +1,0 @@
-/*
- *    uDig - User Friendly Desktop Internet GIS client
- *    http://udig.refractions.net
- *    (C) 2012, Refractions Research Inc.
- *
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * (http://www.eclipse.org/legal/epl-v10.html), and the Refractions BSD
- * License v1.0 (http://udig.refractions.net/files/bsd3-v10.html).
- */
-package org.locationtech.udig.catalog.tests.internal.geotiff;
-
-import org.eclipse.jface.resource.ImageDescriptor;
-import org.eclipse.ui.plugin.AbstractUIPlugin;
-import org.osgi.framework.BundleContext;
-
-/**
- * The main plugin class to be used in the desktop.
- */
-public class GeoTiffPlugin extends AbstractUIPlugin {
-    // The shared instance.
-    private static GeoTiffPlugin plugin;
-
-    /**
-     * The constructor.
-     */
-    public GeoTiffPlugin() {
-        super();
-        plugin = this;
-    }
-
-    /**
-     * This method is called upon plug-in activation
-     */
-    public void start( BundleContext context ) throws Exception {
-        super.start(context);
-    }
-
-    /**
-     * This method is called when the plug-in is stopped
-     */
-    public void stop( BundleContext context ) throws Exception {
-        super.stop(context);
-        plugin = null;
-    }
-
-    /**
-     * Returns the shared instance.
-     */
-    public static GeoTiffPlugin getDefault() {
-        return plugin;
-    }
-
-    /**
-     * Returns an image descriptor for the image file at the given plug-in relative path.
-     * 
-     * @param path the path
-     * @return the image descriptor
-     */
-    public static ImageDescriptor getImageDescriptor( String path ) {
-        return AbstractUIPlugin.imageDescriptorFromPlugin(
-                "org.locationtech.udig.catalog.tests.internal.geotiff", path); //$NON-NLS-1$
-    }
-}
