+++ plugins/org.locationtech.udig.project.tests/build.properties
@@ -1,4 +1,4 @@
-bin.includes = plugin.xml,\
+bin.includes = fragment.xml,\
                test.xml,\
                META-INF/,\
                .,\
