+++ plugins/org.locationtech.udig.project.tests/src/org/locationtech/udig/project/internal/commands/LayerColoursTest.java
@@ -25,7 +25,7 @@ import org.locationtech.udig.core.testsupport.FeatureCreationTestUtil;
 import org.locationtech.udig.project.IMap;
 import org.locationtech.udig.project.internal.Layer;
 import org.locationtech.udig.project.internal.Map;
-import org.locationtech.udig.project.tests.support.AbstractProjectTestCase;
+import org.locationtech.udig.project.testsupport.AbstractProjectTestCase;
 import org.locationtech.udig.project.testsupport.MapTests;
 import org.locationtech.udig.ui.palette.ColourScheme;
