+++ plugins/org.locationtech.udig.project.tests/src/org/locationtech/udig/project/render/TileCalculatorTest.java
@@ -20,7 +20,7 @@ import java.awt.geom.AffineTransform;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.locationtech.udig.project.tests.support.AbstractProjectTestCase;
+import org.locationtech.udig.project.testsupport.AbstractProjectTestCase;
 
 import com.vividsolutions.jts.geom.Envelope;
