+++ plugins/org.locationtech.udig.project.testsupport/META-INF/MANIFEST.MF
@@ -10,6 +10,7 @@ Require-Bundle: org.locationtech.udig.libs,
  org.locationtech.udig.project,
  org.locationtech.udig.style.sld,
  org.locationtech.udig.project.ui,
- org.locationtech.udig.catalog.testsupport
+ org.locationtech.udig.catalog.testsupport,
+ org.junit;bundle-version="[4.0.0,5.0.0)"
 Export-Package: org.locationtech.udig.project.testsupport
 Bundle-ActivationPolicy: lazy
