+++ plugins/org.locationtech.udig.project.testsupport/src/org/locationtech/udig/project/testsupport/AbstractProjectTestCase.java
@@ -8,7 +8,7 @@
  * (http://www.eclipse.org/legal/epl-v10.html), and the Refractions BSD
  * License v1.0 (http://udig.refractions.net/files/bsd3-v10.html).
  */
-package org.locationtech.udig.project.tests.support;
+package org.locationtech.udig.project.testsupport;
 
 import java.io.File;
 import java.util.ArrayList;
