+++ plugins/org.locationtech.udig.render.feature.basic.test/META-INF/MANIFEST.MF
@@ -10,7 +10,6 @@ Require-Bundle: org.eclipse.core.runtime,
  org.locationtech.udig.render.feature.basic,
  org.locationtech.udig.libs,
  org.locationtech.udig.project,
- org.locationtech.udig.project.tests,
  org.locationtech.udig.ui.tests,
  org.locationtech.udig.project.testsupport,
  org.locationtech.udig.core.testsupport
